Adds "Gun" prefab and attaches it to player
I added the "OnFire" input to the StarterAssetsInputs script and created the new "FireGun" script which is attached to a gun model which I made a child of the player capsule.

When you fire it should say "Bang!" in the Debug log, but there is a line you can uncomment in the FireGun script that will also say the name of the object shot in the debug log.

I also added a new models folder which includes several, most likely temporary, low poly gun models we can use for the time being.