Added InputSystem class, gun fire volume slider, and toggleable UI window
Gun fire sound can be changed in the the gun sound script when it's added as a component. Forthe InputSystem (or whatever its called), I just checked a box that added it as a class so we can use it in other scripts for stuff. For the inventory, I made a Scroll View, removed the scroll bars, made a placeholder sprite for it, and made it toggleable with Tab.